requirements.txt.txt
streamlit
ultralytics
pillow
opencv-python
pandas
sqlite3-binary
